Move log into correct domain
