+++ bin/cli/log.js
